+++ recipe/jwt/api/getJWKS.go
@@ -31,8 +31,12 @@ func GetJWKS(apiImplementation jwtmodels.APIInterface, options jwtmodels.APIOpti
 		return err
 	}
 
-	options.Res.Header().Set("Access-Control-Allow-Origin", "*")
-	return supertokens.Send200Response(options.Res, map[string]interface{}{
-		"keys": response.OK.Keys,
-	})
+	if response.OK != nil {
+		options.Res.Header().Set("Access-Control-Allow-Origin", "*")
+		return supertokens.Send200Response(options.Res, map[string]interface{}{
+			"keys": response.OK.Keys,
+		})
+	} else {
+		return supertokens.Send200Response(options.Res, supertokens.ConvertGeneralErrorToJsonResponse(*response.GeneralError))
+	}
 }